+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Quality of life improvements.
 - `BaseScriptableEvent` class which is used internally to draw inspector GUIs. Using this class also improves Odin integration.
 - Support for `Action<TArg>` listeners. This means that regular methods can now be used as listeners without the need of implementing `IScriptableEventListener<TArg>`.
 - Additional _Raise_ button which is shown next to each added listener. Using this button listeners can be raised individually through the inspector.
+- Icons for events and listeners (this will require asset re-import).
 
 ### Changed
 - Renamed Scriptable Event creation menu from _Scriptable **Events**_ to _Scriptable **Event**_.